Refactor 5 - Add _cmponents/__init__.py (#1029)
Summary:
Pull Request resolved: https://github.com/facebookresearch/spdl/pull/1029

Part 5 of https://github.com/facebookresearch/spdl/pull/1026

Add `_components/__init__.py` and import from there.

Differential Revision: D84880870
